deletes old components sets enable pagination default to true
