+++ frontend-nx/apps/edu-hub/components/common/TableGrid/index.tsx
@@ -1,15 +1,6 @@
-import React, { useState, useMemo, useCallback, ReactElement, Dispatch, SetStateAction, useEffect } from 'react';
-import { ApolloError, DocumentNode } from '@apollo/client';
-import {
-  TextField,
-  Checkbox,
-  Select,
-  MenuItem,
-  FormControl,
-  InputLabel,
-  SelectChangeEvent,
-  IconButton,
-} from '@mui/material';
+import { BaseRow, TableGridProps } from './types';
+import React, { useState, useMemo, useCallback } from 'react';
+import { TextField, Checkbox, Select, MenuItem, FormControl, InputLabel, SelectChangeEvent } from '@mui/material';
 import useTranslation from 'next-translate/useTranslation';
 import { ArrowDropUp, ArrowDropDown } from '@mui/icons-material';
 import { MdArrowBack, MdArrowForward, MdDelete } from 'react-icons/md';
@@ -23,135 +14,12 @@ import {
   SortingState,
   useReactTable,
   FilterFn,
-  OnChangeFn,
 } from '@tanstack/react-table';
 import { rankItem } from '@tanstack/match-sorter-utils';
 
-import EhAddButton from '../common/EhAddButton';
-import { QuestionConfirmationDialog } from './dialogs/QuestionConfirmationDialog';
-import { useAdminMutation } from '../../hooks/authedMutation';
-import { useBulkActions } from '../../hooks/bulkActions';
-
-export interface BaseRow {
-  id: number;
-}
-
-export interface BulkAction {
-  value: string;
-  label: string;
-}
-
-interface TableGridDeleteButtonProps {
-  deleteMutation: DocumentNode;
-  id: string | number;
-  refetchQueries: string[];
-  idType: 'number' | 'uuidString';
-  deletionConfirmationQuestion?: string;
-}
-
-const TableGridDeleteButton = ({
-  deleteMutation,
-  id,
-  refetchQueries,
-  idType,
-  deletionConfirmationQuestion,
-}: TableGridDeleteButtonProps) => {
-  const [deleteItem] = useAdminMutation(deleteMutation);
-  const [isConfirmationOpen, setIsConfirmationOpen] = useState(false);
-  const { t } = useTranslation();
-
-  // If no question is provided, use the default one
-  const confirmationQuestion = deletionConfirmationQuestion || t('common:deletion_confirmation_question');
-
-  const handleDeleteClick = () => {
-    setIsConfirmationOpen(true);
-  };
-
-  const handleConfirmationClose = (confirmed: boolean) => {
-    setIsConfirmationOpen(false);
-    if (confirmed) {
-      performDelete();
-    }
-  };
-
-  const performDelete = () => {
-    let variableId: string | number = id;
-
-    if (idType === 'number') {
-      if (typeof id === 'string') {
-        variableId = parseInt(id, 10);
-        if (isNaN(variableId)) {
-          console.error('Invalid numeric ID:', id);
-          return;
-        }
-      }
-    } else if (idType === 'uuidString') {
-      if (typeof id !== 'string') {
-        console.error('Invalid UUID string:', id);
-        return;
-      }
-      // Optionally, you could add a UUID validation regex here
-    }
-
-    deleteItem({
-      variables: { id: variableId },
-      refetchQueries,
-    });
-  };
-
-  return (
-    <>
-      <IconButton
-        size="small"
-        onClick={handleDeleteClick}
-        className="delete-button"
-        sx={{
-          backgroundColor: 'transparent !important',
-          padding: 0,
-          boxShadow: 'none',
-          '&:hover': {
-            backgroundColor: 'rgba(255, 0, 0, 0.1) !important',
-          },
-        }}
-      >
-        <MdDelete size="1.25em" color="red" />
-      </IconButton>
-      <QuestionConfirmationDialog
-        question={confirmationQuestion}
-        confirmationText={t('common:confirm_delete')}
-        open={isConfirmationOpen}
-        onClose={() => handleConfirmationClose(false)}
-        onConfirm={() => handleConfirmationClose(true)}
-      />
-    </>
-  );
-};
-
-interface TableGridProps<T extends BaseRow> {
-  addButtonText?: string;
-  data: T[];
-  columns: ColumnDef<T>[];
-  deleteMutation?: DocumentNode;
-  deleteIdType?: 'number' | 'uuidString';
-  generateDeletionConfirmationQuestion?: (row: T) => string;
-  enablePagination?: boolean;
-  error: ApolloError;
-  expandableRowComponent?: (props: { row: T }) => ReactElement | null;
-  loading: boolean;
-  pageIndex?: number;
-  pages?: number;
-  refetchQueries: string[];
-  searchFilter?: string;
-  setPageIndex?: (number) => void;
-  setSearchFilter?: Dispatch<SetStateAction<string>>;
-  showCheckbox?: boolean;
-  showDelete?: boolean;
-  showGlobalSearchField?: boolean;
-  translationNamespace?: string;
-  onAddButtonClick?: () => void;
-  onBulkAction?: (action: string, selectedRows: T[]) => void;
-  bulkActions?: BulkAction[];
-}
+import EhAddButton from '../EhAddButton';
+import { useBulkActions } from '../../../hooks/bulkActions';
+import TableGridDeleteButton from './components/TableGridDeleteButton';
 
 const TableGrid = <T extends BaseRow>({
   addButtonText,
@@ -160,16 +28,12 @@ const TableGrid = <T extends BaseRow>({
   deleteMutation,
   deleteIdType,
   generateDeletionConfirmationQuestion,
-  enablePagination = false,
+  enablePagination = true,
   error,
   expandableRowComponent,
   loading,
-  pageIndex = 0,
-  pages,
+  pageSize = 15,
   refetchQueries,
-  searchFilter,
-  setSearchFilter,
-  setPageIndex,
   showDelete,
   showGlobalSearchField = true,
   translationNamespace,
@@ -177,6 +41,14 @@ const TableGrid = <T extends BaseRow>({
   onBulkAction,
   bulkActions = [],
 }: TableGridProps<T>) => {
+  const [searchFilter, setSearchFilter] = useState('');
+  const [pageIndex, setPageIndex] = useState(0);
+
+  const onGlobalFilterChange = useCallback((value: string) => {
+    setSearchFilter(value);
+    setPageIndex(0);
+  }, []);
+
   const { t } = useTranslation(translationNamespace);
   const [expandedRows, setExpandedRows] = useState<Set<number>>(new Set());
   const [sorting, setSorting] = useState<SortingState>([]);
@@ -193,11 +65,6 @@ const TableGrid = <T extends BaseRow>({
     isAllSelected,
     isSomeSelected,
   } = useBulkActions<T>(bulkActions, onBulkAction);
-
-  // Add this useEffect to log when bulk actions change
-  useEffect(() => {
-    console.log('Bulk action changed:', bulkAction);
-  }, [bulkAction]);
 
   // Add this new function to handle the Select onChange event
   const handleSelectChange = (event: SelectChangeEvent<string>) => {
@@ -229,14 +96,6 @@ const TableGrid = <T extends BaseRow>({
     addMeta({ itemRank });
     return itemRank.passed;
   };
-
-  const onGlobalFilterChange: OnChangeFn<string> = useCallback(
-    async (value) => {
-      setSearchFilter(value);
-      setPageIndex(0);
-    },
-    [setPageIndex, setSearchFilter]
-  );
 
   const memoizedColumns = useMemo(() => {
     const selectionColumn: ColumnDef<T>[] = showCheckbox
@@ -287,7 +146,7 @@ const TableGrid = <T extends BaseRow>({
     state: {
       sorting,
       globalFilter: searchFilter,
-      ...(enablePagination && { pagination: { pageIndex, pageSize: 15 } }),
+      ...(enablePagination && { pagination: { pageIndex, pageSize } }),
     },
     globalFilterFn: fuzzyFilter,
     onGlobalFilterChange: onGlobalFilterChange,
@@ -300,6 +159,8 @@ const TableGrid = <T extends BaseRow>({
     enableRowSelection: true,
     enableMultiRowSelection: true,
   });
+
+  const totalPages = Math.ceil((data?.length || 0) / pageSize);
 
   return (
     <div>
@@ -486,9 +347,9 @@ const TableGrid = <T extends BaseRow>({
               />
             )}
             <p className="font-medium">
-              {t('common:table_grid.pagination_text', { currentPage: pageIndex + 1, totalPage: pages })}
+              {t('common:table_grid.pagination_text', { currentPage: pageIndex + 1, totalPage: totalPages })}
             </p>
-            {pageIndex < (pages ?? 0) - 1 && (
+            {pageIndex < totalPages - 1 && (
               <MdArrowForward
                 className="border-2 rounded-full cursor-pointer hover:bg-indigo-100"
                 size={30}
